fix: fix compile error related to the duplication of files
