+++ ui/src/main/java/com/baghdad/feature/search/component/RecentlyViewedSection.kt
@@ -1,4 +1,4 @@
-package com.baghdad.ui.feature.search.component
+package com.baghdad.feature.search.component
 
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
