--- ui/src/main/java/com/baghdad/ui/feature/search/component/RecentSearchItem.kt
@@ -1,53 +1,0 @@
-package com.baghdad.ui.feature.search.component
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import com.baghdad.design_system.component.Icon
-import com.baghdad.design_system.component.Text
-import com.baghdad.design_system.theme.Theme
-import com.baghdad.ui.R
-
-@Composable
-fun RecentSearchItem(
-    title: String,
-    onCancelClick: () -> Unit,
-    onRecentSearchClicked: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Row(
-        modifier = modifier
-            .padding(vertical = 12.dp)
-            .clickable { onRecentSearchClicked() },
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Icon(
-            painter = painterResource(id = com.baghdad.design_system.R.drawable.clock),
-            contentDescription = stringResource(id = R.string.clock_icon),
-            tint = Theme.color.hint,
-            modifier = Modifier
-                .padding(end = 8.dp)
-                .size(20.dp)
-        )
-        Text(
-            text = title,
-            style = Theme.typography.body.medium,
-            color = Theme.color.title
-        )
-        Spacer(modifier = Modifier.weight(1f))
-        Icon(
-            painter = painterResource(id = com.baghdad.design_system.R.drawable.cancel),
-            tint = Theme.color.hint,
-            contentDescription = stringResource(id = R.string.cancel_icon),
-            modifier = Modifier.clickable { onCancelClick() })
-    }
-}
-
